feat: add CheckBox component
This commit adds a new CheckBox component to the project. The CheckBox component users to select or deselect a value. It takes in a boolean value and a function to update the value. The component also has a title prop for displaying a label. This feature will trigger aing a MINOR version.