feat: Add new CreateInput component
This commit adds a new CreateInput component that can be used to create new inputs in the application. The component includes a title, subtitle, and close button. It also includes functionality to create new inputs in the dropdown menus for agence, equipe, and poste. The component is used in the Admin page to create new users.